refactor: SectionHeader trailing slot + shared ColorSwatchChip
Add an optional trailing action to SectionHeader and drop the developer
panel's bespoke _ActivityHeader. Consolidate the duplicated accent/avatar
swatch widgets and palette into shared ColorSwatchChip.
