+++ lib/shared/components/section_header.dart
@@ -6,22 +6,34 @@ import 'package:flutter/material.dart';
 /// Consolidates the identical private `_SectionHeader`/`_Header` widgets that
 /// were copy-pasted across the settings, developer, voice and privacy screens.
 class SectionHeader extends StatelessWidget {
-  const SectionHeader(this.title, {super.key});
+  const SectionHeader(this.title, {super.key, this.trailing});
 
   final String title;
+
+  /// Optional widget pinned to the trailing edge of the header (e.g. a "Clear"
+  /// action). When present the title flexes and the right padding tightens.
+  final Widget? trailing;
 
   @override
   Widget build(BuildContext context) {
     final colors = BonfireThemeExtension.of(context);
+    final label = Text(
+      title.toUpperCase(),
+      style: Theme.of(context).textTheme.labelMedium!.copyWith(
+            color: colors.gray,
+            letterSpacing: 0.6,
+          ),
+    );
     return Padding(
-      padding: const EdgeInsets.fromLTRB(16, 12, 16, 4),
-      child: Text(
-        title.toUpperCase(),
-        style: Theme.of(context).textTheme.labelMedium!.copyWith(
-              color: colors.gray,
-              letterSpacing: 0.6,
+      padding: EdgeInsets.fromLTRB(16, 12, trailing != null ? 8 : 16, 4),
+      child: trailing == null
+          ? label
+          : Row(
+              children: [
+                Expanded(child: label),
+                trailing!,
+              ],
             ),
-      ),
     );
   }
 }
